Move and rename informativeTemplate
